경찰 근접 제작
// 이제 탱크가 플레이어를 정확히 바라봄
// 경찰을 근접으로 만들고
// 기존의 경찰 시스템 -> 군인
