+++ Assets/02.Scripts/04.NPC/Enemy/Soilder.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6060f713057aba6499285d5dddc5f9e4
+guid: 95d49e40d9f46be4e837b20d0b58e44d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
