+++ Assets/03.Prefabs/01.Manager/SpawnManager.prefab
@@ -44,12 +44,28 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 37832229269de794bbd8ce539d53cf98, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  stepByScore: e8030000d0070000b80b0000a00f00008813000010270000
   step: 0
   spawnObjectsList:
   - SpawnObjects:
-    - {fileID: 0}
-    isBomber: 1
-    spawnInterval: 10
+    - {fileID: 739172332764853768, guid: 1d805af3516a4de4b8b3b2a3756d0cc6, type: 3}
+    isBomber: 0
+    spawnInterval: 0
+  - SpawnObjects: []
+    isBomber: 0
+    spawnInterval: 0
+  - SpawnObjects: []
+    isBomber: 0
+    spawnInterval: 0
+  - SpawnObjects: []
+    isBomber: 0
+    spawnInterval: 0
+  - SpawnObjects: []
+    isBomber: 0
+    spawnInterval: 0
+  - SpawnObjects: []
+    isBomber: 0
+    spawnInterval: 0
   spawnRadius: 50
   spawnHeight: 2.5
   bomberPrefab: {fileID: 4795624700528817056, guid: 2b7366b25626f0341a381fd86ba6561c,
