+++ Assets/03.Prefabs/03.NPC/Soilder.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bb0b97d12b2a2f7459c8147f88017ecd
+guid: 7c4fd4af40e76a843adb75a51af032ee
 PrefabImporter:
   externalObjects: {}
   userData:
