+++ Assets/03.Prefabs/03.NPC/UFO_01.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bb0b97d12b2a2f7459c8147f88017ecd
+guid: 1d805af3516a4de4b8b3b2a3756d0cc6
 PrefabImporter:
   externalObjects: {}
   userData:
